Complete products feature

+++ products/models.py
@@ -4,5 +4,5 @@ from categories.models import Category
 
 class Product(models.Model):
     name = models.CharField(max_length=255)
-    price = models.DecimalField()
+    price = models.FloatField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
